+++ packages/sdk/jest.setup.js
@@ -1,5 +1,10 @@
 global.FormData = require("form-data")
 global.window = {
 	fetch: require("node-fetch"),
+	dispatchEvent: () => {
+	},
+}
+global.CustomEvent = function CustomEvent() {
+	return
 }
 jest.setTimeout(20000)