+++ packages/sdk/src/config/dev.ts
@@ -1,0 +1,6 @@
+import { Config } from "./type"
+
+export const DEV_CONFIG: Config = {
+	basePath: "https://api-dev.rarible.org",
+	ethereumEnv: "ropsten",
+}