+++ packages/sdk/src/order/bid/domain.ts
@@ -1,0 +1,3 @@
+import { PrepareOrderRequest, PrepareOrderResponse } from "../common"
+
+export type IBid = (request: PrepareOrderRequest) => Promise<PrepareOrderResponse>