+++ packages/sdk/src/order/fill/domain.ts
@@ -1,8 +1,8 @@
-import { Order, OrderId } from "@rarible/api-client"
+import { Order, OrderId, OrderPayout } from "@rarible/api-client"
 import { BigNumber } from "@rarible/types/build/big-number"
-import { OrderPayout } from "@rarible/api-client"
 import { IBlockchainTransaction } from "@rarible/sdk-transaction"
 import { AbstractPrepareResponse } from "../../common/domain"
+import { UnionPart } from "../common"
 
 export type PrepareFillRequest = {
 	/**
@@ -36,18 +36,25 @@ export interface FillRequest {
 	/**
 	 * Origin fees, if not supported by the underlying contract, will throw Error
 	 */
-	originFees?: OrderPayout[]
+	originFees?: UnionPart[]
 	/**
 	 * Payouts, if not supported by the underlying contract, will throw Error
 	 */
-	payouts?: OrderPayout[]
+	payouts?: UnionPart[]
 	/**
 	 * Use infinite approvals (for ERC-20)
 	 */
 	infiniteApproval?: boolean
 }
 
-export interface PrepareFillResponse extends AbstractPrepareResponse<"approve" | "send-tx", FillRequest, IBlockchainTransaction> {
+export type FillActionTypes = "approve" | "send-tx"
+
+export interface PrepareFillResponse
+	extends AbstractPrepareResponse<FillActionTypes, FillRequest, IBlockchainTransaction> {
+	/**
+   * is multiple nft
+   */
+	multiple: boolean
 	/**
 	 * Maximum amount to fill (of NFTs)
 	 */