+++ packages/sdk/src/sdk-blockchains/flow/index.ts
@@ -1,17 +1,26 @@
 import { FlowWallet } from "@rarible/sdk-wallet"
+import { createFlowSdk as createFlowSdkInstance } from "@rarible/flow-sdk"
+import { AuthWithPrivateKey } from "@rarible/flow-sdk/build/types"
 import { IRaribleSdk } from "../../domain"
-import { PrepareMintRequest } from "../../nft/mint/prepare-mint-request.type"
+import { FlowMint } from "./mint"
+import { FlowSell } from "./sell"
+import { FlowBuy } from "./buy"
+import { FlowTransfer } from "./transfer"
+import { FlowBurn } from "./burn"
 
-export function createFlowSdk(wallet: FlowWallet): IRaribleSdk {
-	// const sdk = createRaribleSdk(wallet, options.env, ...)
+export function createFlowSdk(wallet: FlowWallet, auth?: AuthWithPrivateKey): Omit<IRaribleSdk, "apis"> {
+	const sdk = createFlowSdkInstance(wallet.fcl, wallet.network, auth)
 
 	return {
 		nft: {
-			mint: null as any,
+			mint: new FlowMint(sdk).prepare,
+			burn: new FlowBurn(sdk).burn,
+			transfer: new FlowTransfer(sdk).transfer,
 		},
 		order: {
-			fill: null as any,
-			sell: null as any,
+			sell: new FlowSell(sdk, wallet).sell,
+			fill: new FlowBuy(sdk, wallet).buy,
+			bid: null as any,
 		},
 	}
 }