+++ packages/sdk/src/sdk-blockchains/tezos/index.ts
@@ -3,14 +3,17 @@ import { IRaribleSdk } from "../../domain"
 import { Sell } from "./sell"
 import { Fill } from "./fill"
 
-export function createTezosSdk(wallet: TezosWallet): IRaribleSdk {
+export function createTezosSdk(wallet: TezosWallet): Omit<IRaribleSdk, "apis"> {
 	return {
 		nft: {
 			mint: null as any,
+			burn: null as any,
+			transfer: null as any,
 		},
 		order: {
 			fill: new Fill(wallet.provider).fill,
 			sell: new Sell(wallet.provider).sell,
+			bid: null as any,
 		},
 	}
 }